requirements.txt
flask==2.3.3
flask-socketio==5.3.6
opencv-python==4.8.1.78
face-recognition==1.3.0
pillow==10.0.1
numpy==1.24.3
python-socketio==5.9.0
python-engineio==4.7.1
